size_t/ptrdiff_t transition part 3: aso, CDT, libast feature tests (#949)
This is the third of the thickfold patch series, which enables
ksh93 to operate within a 64-bit address space.

The parts of ksh93 affected by this commit are:
- The libast aso library.
- The libast CDT library.
- The libast feature tests.

Change in the number of warnings on Linux when compiling with clang
using -Wsign-compare -Wshorten-64-to-32 -Wsign-conversion
-Wimplicit-int-conversion: 3,178 => 3,088 => 42 (progression from
part 2 => part 3 => part 13)

Progresses https://github.com/ksh93/ksh/issues/592

+++ src/lib/libast/cdt/dtlist.c
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2012 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2025 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2026 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -127,7 +127,7 @@ static void* listat(Dt_t* dt, Dtstat_t* st)
 	{	memset(st, 0, sizeof(Dtstat_t));
 		st->meth  = dt->meth->type;
 		st->size  = dt->data->size;
-		st->space = sizeof(Dtlist_t) + (dt->disc->link >= 0 ? 0 : dt->data->size*sizeof(Dthold_t));
+		st->space = (ssize_t)sizeof(Dtlist_t) + (dt->disc->link >= 0 ? 0 : dt->data->size*(ssize_t)sizeof(Dthold_t));
 	}
 
 	return (void*)dt->data->size;
